Happening place Newai n Happening place, this is item for Happening. Added update delete and edit for this entity, needs more work. Added parent child angular form.

+++ src/app/happenings/happenings.component.ts
@@ -42,8 +42,11 @@ export class HappeningsComponent implements OnInit {
   }
 
   delete(happening: Happening): void {
-    this.happenings = this.happenings.filter(h => h !== happening);
-    this.happeningService.deleteHappening(happening).subscribe();
+
+    if (confirm('Are you sure you want to delete this happening?')) {
+      this.happenings = this.happenings.filter(h => h !== happening);
+      this.happeningService.deleteHappening(happening).subscribe();
+    }
   }
 
 }
